+++ ddl-tpcds/bin_partitioned/web_returns.sql
@@ -5,31 +5,31 @@ drop table if exists web_returns;
 
 create table web_returns
 (
-    wr_returned_time_sk       int,
-    wr_item_sk                int,
-    wr_refunded_customer_sk   int,
-    wr_refunded_cdemo_sk      int,
-    wr_refunded_hdemo_sk      int,
-    wr_refunded_addr_sk       int,
-    wr_returning_customer_sk  int,
-    wr_returning_cdemo_sk     int,
-    wr_returning_hdemo_sk     int,
-    wr_returning_addr_sk      int,
-    wr_web_page_sk            int,
-    wr_reason_sk              int,
+    wr_returned_time_sk       bigint,
+    wr_item_sk                bigint,
+    wr_refunded_customer_sk   bigint,
+    wr_refunded_cdemo_sk      bigint,
+    wr_refunded_hdemo_sk      bigint,
+    wr_refunded_addr_sk       bigint,
+    wr_returning_customer_sk  bigint,
+    wr_returning_cdemo_sk     bigint,
+    wr_returning_hdemo_sk     bigint,
+    wr_returning_addr_sk      bigint,
+    wr_web_page_sk            bigint,
+    wr_reason_sk              bigint,
     wr_order_number           bigint,
     wr_return_quantity        int,
-    wr_return_amt             float,
-    wr_return_tax             float,
-    wr_return_amt_inc_tax     float,
-    wr_fee                    float,
-    wr_return_ship_cost       float,
-    wr_refunded_cash          float,
-    wr_reversed_charge        float,
-    wr_account_credit         float,
-    wr_net_loss               float
+    wr_return_amt             double,
+    wr_return_tax             double,
+    wr_return_amt_inc_tax     double,
+    wr_fee                    double,
+    wr_return_ship_cost       double,
+    wr_refunded_cash          double,
+    wr_reversed_charge        double,
+    wr_account_credit         double,
+    wr_net_loss               double
 )
-partitioned by (wr_returned_date_sk       int)
+partitioned by (wr_returned_date_sk       bigint)
 stored as ${FILE};
 
 insert overwrite table web_returns partition (wr_returned_date_sk)
